+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The Sentry SDK team is happy to announce the immediate availability of Sentry Go
 
 ### Breaking Changes
 
+- Behavioral change for the `MaxBreadcrumbs` client option. Removed the hard limit of 100 breadcrumbs, allowing users to set a larger limit and also changed the default limit from 30 to 100 ([#1106](https://github.com/getsentry/sentry-go/pull/1106)))
+
 - The changes to error handling ([#1075](https://github.com/getsentry/sentry-go/pull/1075)) will affect issue grouping. It is expected that any wrapped and complex errors will be grouped under a new issue group.
 
 ### Features